Make transpiler use etree instead of DOM
To make this easier I've moved it into the parser module, so it can
reuse the namespace stuff defined there.
