+++ docs/source/reading_cellml.rst
@@ -12,5 +12,3 @@ This uses the :class:`cellmlmanip.parser.Parser` class.
 Most users won't need to worry about this, but its documentation is included below for completeness.
 
 .. automodule:: cellmlmanip.parser
-
-.. automodule:: cellmlmanip.transpiler
